requirements.txt
atendentepro==0.6.1
python-dotenv==1.2.1
fastapi==0.110.0
uvicorn==0.24.0.post1
httpx==0.27.0
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

